Update creatures.py and game_funcs.py
Moved rand_step() and sug_step() to Hero class

(I realized that I can change the files by uploading files of the same name, instead of having to edit them through GitHub, oops)
