+++ creatures.py
@@ -1,3 +1,4 @@
+import random as ra
 
 class Hero:
 
@@ -5,9 +6,56 @@ class Hero:
     # FIXME: maybe add a gear varaible? not sure how to handle that
 
     symbol = "☺"
+    # These values should be treated as placeholders for now, and later we'll allow the player to create a character.
     health = 100
     attack = 5
     defense = 10
+
+    def rand_step(posx, posy, scale):
+        next_move = ra.choice(["up", "down", "left", "right", "nothing"])
+        if next_move == "up":
+            return posx, posy + scale
+        if next_move == "down":
+            return posx, posy - scale
+        if next_move == "left":
+            return posx - scale, posy
+        if next_move == "right":
+            return posx + scale, posy
+        if next_move == "nothing":
+            return posx, posy
+
+    def sug_step(posx, posy, sugx, sugy, sugtype, scale):
+
+        up, down, left, right = False, False, False, False
+
+        if sugtype == "MoveTo":
+            if posx > sugx:
+                left = True
+            elif posx < sugx:
+                right = True
+            if posy > sugy:
+                up = True
+            elif posy < sugy:
+                down = True
+
+            if right and up:
+                return posx + scale, posy - scale
+            elif left and up:
+                return posx - scale, posy - scale
+            elif right and down:
+                return posx + scale, posy + scale
+            elif left and down:
+                return posx - scale, posy + scale
+            elif left:
+                return posx - scale, posy
+            elif right:
+                return posx + scale, posy
+            elif up:
+                return posx, posy - scale
+            elif down:
+                return posx, posy + scale
+            else:
+                return posx, posy
 
 
 class Goblin:
